refactor: Move useAccountIconInitial(), useStoreLoginState(), and useLogOut().
