+++ app/src/resources/access-control/useAccountIconInitial.ts
